+++ src/main.c
@@ -1,0 +1,6 @@
+#include "nd2tool.h"
+
+int main(int argc, char ** argv)
+{
+    return nd2tool_cli(argc, argv);
+}